+++ app/src/main/java/com/example/teashop/data/model/product/ProductShort.kt
@@ -1,13 +1,10 @@
 package com.example.teashop.data.model.product
 
 import android.os.Parcelable
-import com.example.teashop.data.model.category.Category
 import com.example.teashop.data.model.image.Image
-import com.example.teashop.data.model.packages.PackageProduct
 import com.example.teashop.data.model.packages.PackageShort
-import com.example.teashop.data.model.packages.ShortOrderPackage
 import kotlinx.parcelize.Parcelize
-import java.time.ZonedDateTime
+
 @Parcelize
 data class ProductShort(
     val id: Long = 1234,
@@ -15,7 +12,7 @@ data class ProductShort(
     val isFavorite: Boolean = false,
     val packages: List<PackageShort> = listOf(PackageShort()),
     val title: String = "Зелёный чай",
-    val discount: Double = 5.0,
+    val discount: Int = 5,
     val countOfReviews: Int = 6,
     val averageRating: Double = 4.9,
     val images: List<Image> = listOf(Image())
